target controller
ClusterTargets and Targets CRDs allow the user to add cluster-scoped and namespaced targets at runtime. They replace the configuration file (in a config map) built by Helm at install/upgrade time. When the controller-manager and proxy-scheduler start, they list targets, get the corresponding secrets, and create client configs from them. No need to mount the secrets.

Since adding informers to running controllers isn't straightforward, we simply watch target objects in the new restarter deployment and restart the controller-manager and proxy-scheduler when targets change.

We removed the configuration file entirely for now, therefore the ability to override the webhook address, which was only useful to redirect webhook calls to the developer's machine while debugging, was removed (not an end-user facing feature so that's okay).

The main README is restructured to first install multicluster-scheduler in all clusters (identically), then connect sources and targets (which can be repeated when adding targets to existing sources). I added alternative instructions for people not wanting to use klum or kubemcsa.

Redundant parts of the Helm chart's documentation have been removed to consolidate installation instructions in the main README. What remains is a discussion about the post-delete hook, and the customary description of Helm values.

Signed-off-by: adrienjt <adrienjt@users.noreply.github.com>