refactor controllers from fan-out to 1-on-1
The problem with fan-out controllers was that they wouldn't start if one target was unavailable (blocked at WaitForCacheSync, and we'd rather sync the cache before reconciling).

Another benefit is the use of more namespaced informers (safer, faster).

Note: this reopens a rare race condition where candidates can be orphaned if proxy pods are deleted before finalizers are added (they can't be added before because that would invalidate the proxy scheduler cache mid-cycle). This is a trade-off. We will eventually need proper GC and let go of finalizers and their many flaws.
Signed-off-by: adrienjt <adrienjt@users.noreply.github.com>